+++ app/src/main/java/dev/vanilson/jamma/transaction/domain/Wallet.kt
@@ -3,5 +3,5 @@ package dev.vanilson.jamma.transaction.domain
 data class Wallet(
     val uid: Int,
     val name: String,
-    val balance: Double = 0.00,
+    val balance: Long = 0L,
 )
